refactor(repository): MessageRepository를 인터페이스로 만들고 JCFMessageRepository 클래스를 생성
- 원래 작성했던 MessageRepository클래스를 JCFMessageRepository로 옮김
